Removed two of three eval commands
The eval on line 41 and 44 weren't needed. I kept the one on line 50 as removing it gives me an error when running the script and doesn't execute (or remove the tarball on the reMarkable device). Error was:
sh: cd /usr/share/remarkable/templates/ && tar -xf archivedTemplates.tgz && rm archivedTemplates.tgz: No such file or directory