Create compress_and_watermark_my_images.sh
Uploading my working script that takes all the images in a directory, where the script is executed, and then moves, compresses, renames, and watermarks each image before returning them to the original directory and removing the temporary one. The full pathway of the watermark image is hardcoded within the script. If you want to dictate the directory and image when running the script you'll want to look at the related one: compress_and_watermark_my_images_with_arguments.sh