Backup your reMarkable
Just created a quick script to backup my reMarkable. I put this in my backup directory. When run it creates a dated folder and executes several scp commands to back up the files on the reMarkable. Before using it replace your IP address on line 5 and, if needed, setup a public key so you don't need to physical enter a password when connecting to your device.